requirements.txt
aiogram>=3.14.0
asyncpg>=0.30.0
SQLAlchemy>=2.0.37
APScheduler>=3.10.4
requests>=2.31.0
attrs>=23.2.0
python-dotenv>=1.0.1
